+++ src/main/main.ts
@@ -100,6 +100,7 @@ async function createWindow(): Promise<BrowserWindow> {
     height: 728,
     icon: getAssetPath('icon.png'),
     webPreferences: {
+      webSecurity: false,
       preload: app.isPackaged
         ? path.join(__dirname, 'preload.js')
         : path.join(__dirname, '../../.erb/dll/preload.js'),